backend/requirements.txt
langgraph>=0.0.10
python-dotenv>=1.0.0
grpcio>=1.60.0
flask>=2.3.3
flask-cors>=4.0.0
llama-index>=0.14.3
sqlalchemy>=2.0.43
langchain_google_genai>=2.1.12
llama-index.llms.gemini>=0.6.1
pymysql>=1.0.2